chore: cleanup and make it build
