+++ apps/web/src/components/accounts/liquidity/new-account.tsx
@@ -1,0 +1,7 @@
+export default function NewAccount() {
+  return (
+    <div>
+      <div>New Liquidity</div>
+    </div>
+  );
+}
