+++ apps/web/src/config/account.ts
@@ -1,35 +1,34 @@
 import { AccountType } from "@/gen/proto/fijoy/v1/account_pb";
+import { AccountTypeDetail } from "@/types/account";
+import { match } from "ts-pattern";
 
-type AccountTypeConfig = {
-  isDebt: boolean;
-  name: string;
-  description: string;
-};
-
-export const accountTypeConfigMap: Record<
-  Exclude<AccountType, AccountType.UNSPECIFIED>,
-  AccountTypeConfig
-> = {
-  [AccountType.CASH]: {
-    isDebt: false,
-    name: "Cash",
-    description: "Holds liquid funds for immediate use.",
-  },
-  [AccountType.DEBT]: {
-    isDebt: true,
-    name: "Debt",
-    description: "Represents money owed, such as loans or credit balances.",
-  },
-  [AccountType.INVESTMENT]: {
-    isDebt: false,
-    name: "Investment",
-    description:
-      "Used for buying and selling securities like stocks and bonds.",
-  },
-  [AccountType.OTHER_ASSET]: {
-    isDebt: false,
-    name: "Other Asset",
-    description:
-      "Includes valuable possessions like real estate, vehicles, or collectibles.",
-  },
-};
+export function getAccountTypeDetail(
+  accountType: AccountType,
+): AccountTypeDetail {
+  return match(accountType)
+    .with(AccountType.UNSPECIFIED as 0, () => ({
+      name: "Unspecified",
+      description: "Unspecified",
+    }))
+    .with(AccountType.LIQUIDITY as 1, () => ({
+      name: "Liquidity",
+      description: "Liquid assets that can be quickly converted to cash",
+    }))
+    .with(AccountType.INVESTMENT as 2, () => ({
+      name: "Investment",
+      description: "Investment assets such as stocks, cryptocurrencies, etc.",
+    }))
+    .with(AccountType.PROPERTY as 3, () => ({
+      name: "Property",
+      description: "Property assets such as real estate, vehicles, etc.",
+    }))
+    .with(AccountType.RECEIVABLE as 4, () => ({
+      name: "Receivable",
+      description: "Assets that other people owe to you",
+    }))
+    .with(AccountType.LIABILITY as 5, () => ({
+      name: "Liability",
+      description: "Debts such as loans, mortgages, etc.",
+    }))
+    .exhaustive();
+}
