+++ apps/web/src/routes/_protected/_profile/transactions/index.tsx
@@ -8,11 +8,9 @@ import {
 } from "@/components/small-header";
 // import { accountsQueryOptions } from "@/lib/queries/account";
 // import { categoriesQueryOptions } from "@/lib/queries/category";
-import { useSuspenseQuery } from "@tanstack/react-query";
+// import { useSuspenseQuery } from "@tanstack/react-query";
 import { createFileRoute } from "@tanstack/react-router";
 // import { getAccountsQueryOptions } from "@/lib/queries/account";
-import { DataTable } from "@/components/transactions/data-table";
-import { columns } from "@/components/transactions/columns";
 import CenterLoadingSpinner from "@/components/center-loading-spinner";
 
 export const Route = createFileRoute("/_protected/_profile/transactions/")({
@@ -25,17 +23,17 @@ export const Route = createFileRoute("/_protected/_profile/transactions/")({
 });
 
 function Page() {
-  const context = Route.useRouteContext();
+  // const context = Route.useRouteContext();
 
   // const { data: categories } = useSuspenseQuery(
   //   categoriesQueryOptions(workspace.id),
   // );
 
-  const transactionsQuery = useSuspenseQuery(
-    getTransactionsQueryOptions({ context }),
-  );
+  // const transactionsQuery = useSuspenseQuery(
+  //   getTransactionsQueryOptions({ context }),
+  // );
 
-  const transactions = transactionsQuery.data.transactions;
+  // const transactions = transactionsQuery.data.transactions;
 
   // const accountsQuery = useSuspenseQuery(getAccountsQueryOptions({ context }));
 
@@ -59,8 +57,6 @@ function Page() {
       {/*   workspace={context.workspace} */}
       {/*   categories={[]} // FIXME: load categories */}
       {/* /> */}
-
-      <DataTable columns={columns} data={transactions} />
     </>
   );
 }
